IOS-7018 Folders grooming
