--- Tangem/Modules/Send/Fee/CustomFeeService/CustomFeeServiceInputOutput.swift
@@ -1,10 +1,0 @@
-//
-//  CustomFeeServiceInputOutput.swift
-//  Tangem
-//
-//  Created by Andrey Chukavin on 02.04.2024.
-//  Copyright © 2024 Tangem AG. All rights reserved.
-//
-
-import BlockchainSdk
-import Combine
