+++ Tangem/Modules/Send/Fee/SendFeeViewModel.swift
@@ -9,201 +9,7 @@
 import Foundation
 import SwiftUI
 import Combine
-import BigInt
 import BlockchainSdk
-
-class CommonSendFeeProvider: SendFeeProvider {
-    private let walletModel: WalletModel
-
-    init(walletModel: WalletModel) {
-        self.walletModel = walletModel
-    }
-
-    func getFee(amount: Amount, destination: String) -> AnyPublisher<[Fee], any Error> {
-        walletModel.getFee(amount: amount, destination: destination)
-    }
-}
-
-protocol SendFeeProvider {
-    func getFee(amount: Amount, destination: String) -> AnyPublisher<[Fee], Error>
-}
-
-struct SendFee: Hashable {
-    let option: FeeOption
-    let value: LoadingValue<Fee>
-}
-
-class CommonSendFeeProcessor {
-    private let provider: SendFeeProvider
-    private var customFeeService: CustomFeeService?
-
-    private let _cryptoAmount: CurrentValueSubject<Amount?, Never> = .init(.none)
-    private let _destination: CurrentValueSubject<String?, Never> = .init(.none)
-    private let _fees: CurrentValueSubject<[SendFee], Never> = .init([])
-
-    private let _customFee: CurrentValueSubject<SendFee?, Never> = .init(.none)
-    private let defaultFeeOptions: [FeeOption]
-    private var feeOptions: [FeeOption] {
-        var options = defaultFeeOptions
-        if supportCustomFee {
-            options.append(.custom)
-        }
-        return options
-    }
-
-    private var supportCustomFee: Bool {
-        customFeeService != nil
-    }
-
-    private var bag: Set<AnyCancellable> = []
-
-    init(
-        provider: SendFeeProvider,
-        defaultFeeOptions: [FeeOption],
-        customFeeServiceFactory: CustomFeeServiceFactory
-    ) {
-        self.provider = provider
-        self.defaultFeeOptions = defaultFeeOptions
-
-        customFeeService = customFeeServiceFactory.makeService(input: self, output: self)
-    }
-
-    func bind(input: SendFeeProcessorInput) {
-        input.cryptoAmountPublisher
-            .withWeakCaptureOf(self)
-            .sink { processor, amount in
-                processor._cryptoAmount.send(amount)
-            }
-            .store(in: &bag)
-
-        input.destinationPublisher
-            .withWeakCaptureOf(self)
-            .sink { processor, destination in
-                processor._destination.send(destination)
-            }
-            .store(in: &bag)
-    }
-}
-
-// MARK: - CustomFeeServiceInput
-
-extension CommonSendFeeProcessor: CustomFeeServiceInput {
-    var cryptoAmountPublisher: AnyPublisher<Amount, Never> {
-        _cryptoAmount.compactMap { $0 }.eraseToAnyPublisher()
-    }
-
-    var destinationPublisher: AnyPublisher<String, Never> {
-        _destination.compactMap { $0 }.eraseToAnyPublisher()
-    }
-}
-
-// MARK: - CustomFeeServiceOutput
-
-extension CommonSendFeeProcessor: CustomFeeServiceOutput {
-    func customFeeDidChanged(_ customFee: Fee?) {
-        let fee = customFee.map { SendFee(option: .custom, value: .loaded($0)) }
-        _customFee.send(fee)
-    }
-}
-
-// MARK: - SendFeeProcessor
-
-extension CommonSendFeeProcessor: SendFeeProcessor {
-    func setup(input: SendFeeProcessorInput) {
-        bind(input: input)
-    }
-
-    func updateFees() {
-        guard let amount = _cryptoAmount.value,
-              let destination = _destination.value else {
-            assertionFailure("SendFeeProcessor is not ready to update fees")
-            return
-        }
-
-        provider
-            .getFee(amount: amount, destination: destination)
-            .sink(receiveCompletion: { [weak self] completion in
-                guard case .failure(let error) = completion else {
-                    return
-                }
-
-                self?.update(fees: .failedToLoad(error: error))
-            }, receiveValue: { [weak self] fees in
-                self?.update(fees: .loaded(fees))
-            })
-            .store(in: &bag)
-    }
-
-    func feesPublisher() -> AnyPublisher<[SendFee], Never> {
-        _fees.dropFirst().eraseToAnyPublisher()
-    }
-
-    func customFeeInputFieldModels() -> [SendCustomFeeInputFieldModel] {
-        customFeeService?.inputFieldModels() ?? []
-    }
-}
-
-// MARK: - Private
-
-private extension CommonSendFeeProcessor {
-    func update(fees value: LoadingValue<[Fee]>) {
-        switch value {
-        case .loading:
-            _fees.send(feeOptions.map { SendFee(option: $0, value: .loading) })
-        case .loaded(let fees):
-            _fees.send(mapToFees(fees: fees))
-        case .failedToLoad(let error):
-            _fees.send(feeOptions.map { SendFee(option: $0, value: .failedToLoad(error: error)) })
-        }
-    }
-
-    func mapToFees(fees: [Fee]) -> [SendFee] {
-        var defaultOptions = mapToDefaultFees(fees: fees)
-
-        if supportCustomFee {
-            var customFee = _customFee.value
-
-            if customFee == nil, let market = defaultOptions.first(where: { $0.option == .market }) {
-                customFee = SendFee(option: .custom, value: market.value)
-            }
-
-            if let custom = customFee {
-                defaultOptions.append(custom)
-            }
-        }
-
-        return defaultOptions
-    }
-
-    func mapToDefaultFees(fees: [Fee]) -> [SendFee] {
-        switch fees.count {
-        case 1:
-            return [SendFee(option: .market, value: .loaded(fees[1]))]
-        case 3:
-            return [
-                SendFee(option: .slow, value: .loaded(fees[0])),
-                SendFee(option: .market, value: .loaded(fees[1])),
-                SendFee(option: .fast, value: .loaded(fees[2])),
-            ]
-        default:
-            assertionFailure("Wrong count of fees")
-            return []
-        }
-    }
-}
-
-protocol SendFeeProcessorInput: AnyObject {
-    var cryptoAmountPublisher: AnyPublisher<Amount, Never> { get }
-    var destinationPublisher: AnyPublisher<String, Never> { get }
-}
-
-protocol SendFeeProcessor {
-    func updateFees()
-    func feesPublisher() -> AnyPublisher<[SendFee], Never>
-    func customFeeInputFieldModels() -> [SendCustomFeeInputFieldModel]
-
-    func setup(input: SendFeeProcessorInput)
-}
 
 protocol SendFeeInput: AnyObject {
     var selectedFee: SendFee? { get }
@@ -242,13 +48,6 @@ class SendFeeViewModel: ObservableObject {
     private let processor: SendFeeProcessor
     private let notificationManager: SendNotificationManager
 
-//    private let walletInfo: SendWalletInfo
-//    private let customFeeService: CustomFeeService?
-//    private let customFeeInFiat = CurrentValueSubject<String?, Never>("")
-    // Save this values to compare it when the focus changed and send analytics
-//    private var customFeeValue: Decimal?
-//    private var customFeeBeforeEditing: Decimal?
-
     private let feeExplanationUrl = TangemBlogUrlBuilder().url(post: .fee)
     private let balanceFormatter = BalanceFormatter()
     private let balanceConverter = BalanceConverter()
@@ -277,18 +76,6 @@ class SendFeeViewModel: ObservableObject {
         self.processor = processor
         self.notificationManager = notificationManager
 
-//        self.customFeeService = customFeeService
-//        self.walletInfo = walletInfo
-//        feeOptions = input.feeOptions
-//        selectedFeeOption = input.selectedFeeOption
-
-//        if feeOptions.contains(.custom) {
-//            createCustomFeeModels()
-//        }
-
-//        feeRowViewModels = makeFeeRowViewModels([:])
-
-//        setupView()
         bind()
     }
 
@@ -315,42 +102,6 @@ class SendFeeViewModel: ObservableObject {
         router?.openFeeExplanation(url: feeExplanationUrl)
     }
 
-    /*
-     private func createCustomFeeModels() {
-         guard let customFeeService else { return }
-
-         let editableCustomFeeService = customFeeService as? EditableCustomFeeService
-         let onCustomFeeFieldChange: ((Decimal?) -> Void)?
-         if let editableCustomFeeService {
-             onCustomFeeFieldChange = { [weak self] value in
-                 self?.customFeeValue = value
-                 editableCustomFeeService.setCustomFee(value: value)
-             }
-         } else {
-             onCustomFeeFieldChange = nil
-         }
-
-         let customFeeModel = SendCustomFeeInputFieldModel(
-             title: Localization.sendMaxFee,
-             amountPublisher: input.customFeePublisher.decimalPublisher,
-             disabled: editableCustomFeeService == nil,
-             fieldSuffix: walletInfo.feeCurrencySymbol,
-             fractionDigits: walletInfo.feeFractionDigits,
-             amountAlternativePublisher: customFeeInFiat.eraseToAnyPublisher(),
-             footer: customFeeService.customFeeDescription,
-             onFieldChange: onCustomFeeFieldChange
-         ) { [weak self] focused in
-             self?.onCustomFeeChanged(focused)
-         }
-
-         customFeeModels = [customFeeModel] + customFeeService.inputFieldModels()
-     }
-     */
-
-//    private func setupView() {
-//        updateViewModels(values: feeOptions.reduce(into: [:]) { $0[$1] = .loading })
-//    }
-
     private func bind() {
         processor.feesPublisher()
             .withWeakCaptureOf(self)
@@ -368,30 +119,6 @@ class SendFeeViewModel: ObservableObject {
                 viewModel.updateSelectedOption(selectedFee: selectedFee)
             }
             .store(in: &bag)
-
-//        input
-//            .customFeePublisher
-//            .withWeakCaptureOf(self)
-//            .map { (self, customFee) -> String? in
-//                guard
-//                    let customFee,
-//                    let feeCurrencyId = self.walletInfo.feeCurrencyId,
-//                    let fiatFee = self.balanceConverter.convertToFiat(customFee.amount.value, currencyId: feeCurrencyId)
-//                else {
-//                    return nil
-//                }
-//
-//                return self.balanceFormatter.formatFiatBalance(fiatFee)
-//            }
-//            .withWeakCaptureOf(self)
-//            .sink { (self, customFeeInFiat) in
-//                self.customFeeInFiat.send(customFeeInFiat)
-//            }
-//            .store(in: &bag)
-
-//        input.isFeeIncludedPublisher
-//            .assign(to: \.isFeeIncluded, on: self, ownership: .weak)
-//            .store(in: &bag)
 
         notificationManager
             .notificationPublisher(for: .feeLevels)
@@ -475,19 +202,5 @@ extension SendFeeViewModel {
     struct Initial {
         let tokenItem: TokenItem
         let feeOptions: [FeeOption]
-    }
-}
-
-// MARK: - private extensions
-
-private extension AnyPublisher where Output == Fee?, Failure == Never {
-    var decimalPublisher: AnyPublisher<Decimal?, Never> {
-        map { $0?.amount.value }.eraseToAnyPublisher()
-    }
-}
-
-private extension AnyPublisher where Output == BigUInt?, Failure == Never {
-    var decimalPublisher: AnyPublisher<Decimal?, Never> {
-        map { $0?.decimal }.eraseToAnyPublisher()
     }
 }
