+++ Tangem/Modules/Send/Models/SendFee.swift
@@ -1,0 +1,14 @@
+//
+//  SendFee.swift
+//  Tangem
+//
+//  Created by Sergey Balashov on 18.06.2024.
+//  Copyright © 2024 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+struct SendFee: Hashable {
+    let option: FeeOption
+    let value: LoadingValue<Fee>
+}
